♻️ Separate `ui_components.py`
