+++ food_classifier/src/service_ui/clients/db_client.py
@@ -80,6 +80,7 @@ class DatabaseClient:
     def get_food_info_from_db(self, food_name):
         """
         Query the nutrition database for food information based on the food name.
+        Also includes timestamp of when the food was consumed (created_at).
         """
         if not self.connection:
             print("No database connection.")
@@ -88,12 +89,17 @@ class DatabaseClient:
         try:
             cursor = self.connection.cursor(dictionary=True)
             
-            # Query for food information
-            cursor.execute("SELECT * FROM nutrition_info WHERE food_name = %s", (food_name,))
+            # Query for food information and created_at as consumption time
+            cursor.execute("""
+                SELECT n.*, cd.created_at as consumption_time 
+                FROM nutrition_info n 
+                LEFT JOIN customer_diets cd ON n.food_name = cd.food_name 
+                WHERE n.food_name = %s
+                ORDER BY cd.created_at DESC LIMIT 1
+            """, (food_name,))
             food_info = cursor.fetchone()
             
             cursor.close()
-            
             return food_info
             
         except mysql.connector.Error as err:
